refactor: Deduplicate code in restake
